feat: export keycloak realm on deploy
This commit adds a step to export the Keycloak realm during the deployment process.

The `export-realm.sh` script is executed after pulling the latest changes and before stopping the existing containers. This ensures that the Keycloak realm configuration is backed up before any changes are made to the running containers. This provides a backup of the keycloak realm configuration.